+++ apps/ui/src/components/shared/index.ts
@@ -5,3 +5,17 @@ export {
   type UseModelOverrideOptions,
   type UseModelOverrideResult,
 } from './use-model-override';
+
+// Onboarding Wizard Components
+export {
+  OnboardingWizard,
+  useOnboardingWizard,
+  ONBOARDING_STORAGE_PREFIX,
+  ONBOARDING_TARGET_ATTRIBUTE,
+  ONBOARDING_ANALYTICS,
+  type OnboardingStep,
+  type OnboardingState,
+  type OnboardingWizardProps,
+  type UseOnboardingWizardOptions,
+  type UseOnboardingWizardResult,
+} from './onboarding';
