+++ apps/ui/src/components/views/board-view/hooks/use-board-onboarding.ts
@@ -1,19 +1,23 @@
 /**
  * Board Onboarding Hook
  *
- * Manages the state and logic for the interactive onboarding wizard
- * that guides new users through the Kanban board workflow.
+ * Board-specific wrapper around the shared onboarding wizard hook.
+ * Manages board-specific features like sample data (Quick Start).
  *
- * Features:
- * - Tracks wizard completion status per project
- * - Persists state to localStorage (per user, per board)
- * - Handles step navigation
- * - Provides analytics tracking
+ * Usage:
+ * - Wizard is triggered manually via startWizard() when user clicks help button
+ * - No auto-show logic - user controls when to see the wizard
  */
 
-import { useState, useCallback, useEffect, useMemo } from 'react';
+import { useState, useCallback, useEffect } from 'react';
 import { createLogger } from '@automaker/utils/logger';
 import { getItem, setItem } from '@/lib/storage';
+import {
+  useOnboardingWizard,
+  ONBOARDING_TARGET_ATTRIBUTE,
+  type OnboardingStep,
+} from '@/components/shared/onboarding';
+import { PlayCircle, Sparkles, Lightbulb, CheckCircle2, Settings2 } from 'lucide-react';
 
 const logger = createLogger('BoardOnboarding');
 
@@ -21,389 +25,251 @@ const logger = createLogger('BoardOnboarding');
 // CONSTANTS
 // ============================================================================
 
-/** Storage key prefix for onboarding state */
-const ONBOARDING_STORAGE_KEY = 'automaker:board-onboarding';
-
-/** Delay before auto-showing wizard to let the board render first (ms) */
-const WIZARD_AUTO_SHOW_DELAY_MS = 500;
+/** Storage key prefix for board-specific onboarding data */
+const BOARD_ONBOARDING_STORAGE_KEY = 'automaker:board-onboarding-data';
 
 /** Maximum length for project path hash in storage key */
 const PROJECT_PATH_HASH_MAX_LENGTH = 50;
 
-// Analytics event names
-export const ONBOARDING_ANALYTICS = {
-  STARTED: 'onboarding_started',
-  COMPLETED: 'onboarding_completed',
-  SKIPPED: 'onboarding_skipped',
-  QUICK_START_USED: 'onboarding_quick_start_used',
-  SAMPLE_DATA_CLEARED: 'onboarding_sample_data_cleared',
-  STEP_VIEWED: 'onboarding_step_viewed',
-  RETRIGGERED: 'onboarding_retriggered',
+// Board-specific analytics events
+export const BOARD_ONBOARDING_ANALYTICS = {
+  QUICK_START_USED: 'board_onboarding_quick_start_used',
+  SAMPLE_DATA_CLEARED: 'board_onboarding_sample_data_cleared',
 } as const;
 
-// Wizard step definitions
-export interface WizardStep {
-  id: string;
-  columnId: string;
-  title: string;
-  description: string;
-  tip?: string;
-}
+// ============================================================================
+// WIZARD STEPS
+// ============================================================================
 
-export const WIZARD_STEPS: WizardStep[] = [
+/**
+ * Board wizard step definitions
+ * Each step targets a kanban column via data-onboarding-target
+ */
+export const BOARD_WIZARD_STEPS: OnboardingStep[] = [
   {
     id: 'backlog',
-    columnId: 'backlog',
+    targetId: 'backlog',
     title: 'Backlog',
     description:
       'This is where all your planned tasks live. Add new features, bug fixes, or improvements here. When you\'re ready to work on something, drag it to "In Progress" or click the play button.',
     tip: 'Press N or click the + button to quickly add a new feature.',
+    icon: PlayCircle,
   },
   {
     id: 'in_progress',
-    columnId: 'in_progress',
+    targetId: 'in_progress',
     title: 'In Progress',
     description:
       'Tasks being actively worked on appear here. AI agents automatically pick up items from the backlog and move them here when processing begins.',
     tip: 'You can run multiple tasks simultaneously using Auto Mode.',
+    icon: Sparkles,
   },
   {
     id: 'waiting_approval',
-    columnId: 'waiting_approval',
+    targetId: 'waiting_approval',
     title: 'Waiting Approval',
     description:
       'Completed work lands here for your review. Check the changes, run tests, and approve or send back for revisions.',
     tip: 'Click "View Output" to see what the AI agent did.',
+    icon: Lightbulb,
   },
   {
     id: 'verified',
-    columnId: 'verified',
+    targetId: 'verified',
     title: 'Verified',
     description:
       "Approved and verified tasks are ready for deployment! Archive them when you're done or move them back if changes are needed.",
     tip: 'Click "Complete All" to archive all verified items at once.',
+    icon: CheckCircle2,
   },
   {
     id: 'custom_columns',
-    columnId: 'in_progress', // Highlight "In Progress" column to show the settings icon
+    targetId: 'pipeline-settings', // Highlight the pipeline settings button icon
     title: 'Custom Pipelines',
     description:
-      'You can create custom columns (called pipelines) to build your own workflow! Click the settings icon in any column header to add, rename, or configure pipeline steps.',
+      'You can create custom columns (called pipelines) to build your own workflow! Click this settings icon to add, rename, or configure pipeline steps.',
     tip: 'Use pipelines to add code review, QA testing, or any custom stage to your workflow.',
+    icon: Settings2,
   },
 ];
 
-// Persisted onboarding state structure
-interface OnboardingState {
-  completed: boolean;
-  completedAt?: string;
-  skipped: boolean;
-  skippedAt?: string;
-  hasEverSeenWizard: boolean;
+// Re-export for backward compatibility
+export type { OnboardingStep as WizardStep } from '@/components/shared/onboarding';
+export { ONBOARDING_TARGET_ATTRIBUTE };
+
+// ============================================================================
+// BOARD-SPECIFIC STATE
+// ============================================================================
+
+interface BoardOnboardingData {
   hasSampleData: boolean;
   quickStartUsed: boolean;
 }
 
-// Default state for new projects
-const DEFAULT_ONBOARDING_STATE: OnboardingState = {
-  completed: false,
-  skipped: false,
-  hasEverSeenWizard: false,
+const DEFAULT_BOARD_DATA: BoardOnboardingData = {
   hasSampleData: false,
   quickStartUsed: false,
 };
 
 /**
- * Get storage key for a specific project
- * Creates a sanitized key from the project path for localStorage
+ * Sanitize project path to create a storage key
  */
-function getStorageKey(projectPath: string): string {
-  // Create a simple hash of the project path to use as key
-  const hash = projectPath.replace(/[^a-zA-Z0-9]/g, '_').slice(0, PROJECT_PATH_HASH_MAX_LENGTH);
-  return `${ONBOARDING_STORAGE_KEY}:${hash}`;
+function sanitizeProjectPath(projectPath: string): string {
+  return projectPath.replace(/[^a-zA-Z0-9]/g, '_').slice(0, PROJECT_PATH_HASH_MAX_LENGTH);
 }
 
-// Load onboarding state from localStorage
-function loadOnboardingState(projectPath: string): OnboardingState {
+/**
+ * Get storage key for board-specific data
+ */
+function getBoardDataStorageKey(projectPath: string): string {
+  const hash = sanitizeProjectPath(projectPath);
+  return `${BOARD_ONBOARDING_STORAGE_KEY}:${hash}`;
+}
+
+/**
+ * Load board-specific onboarding data from localStorage
+ */
+function loadBoardData(projectPath: string): BoardOnboardingData {
   try {
-    const key = getStorageKey(projectPath);
+    const key = getBoardDataStorageKey(projectPath);
     const stored = getItem(key);
     if (stored) {
-      return JSON.parse(stored) as OnboardingState;
+      return JSON.parse(stored) as BoardOnboardingData;
     }
   } catch (error) {
-    logger.error('Failed to load onboarding state:', error);
+    logger.error('Failed to load board onboarding data:', error);
   }
-  return { ...DEFAULT_ONBOARDING_STATE };
+  return { ...DEFAULT_BOARD_DATA };
 }
 
-// Save onboarding state to localStorage
-function saveOnboardingState(projectPath: string, state: OnboardingState): void {
+/**
+ * Save board-specific onboarding data to localStorage
+ */
+function saveBoardData(projectPath: string, data: BoardOnboardingData): void {
   try {
-    const key = getStorageKey(projectPath);
-    setItem(key, JSON.stringify(state));
+    const key = getBoardDataStorageKey(projectPath);
+    setItem(key, JSON.stringify(data));
   } catch (error) {
-    logger.error('Failed to save onboarding state:', error);
+    logger.error('Failed to save board onboarding data:', error);
   }
 }
 
-// Track analytics event (placeholder - integrate with actual analytics service)
+/**
+ * Track analytics event (placeholder)
+ */
 function trackAnalytics(event: string, data?: Record<string, unknown>): void {
   logger.debug(`[Analytics] ${event}`, data);
-  // TODO: Integrate with actual analytics service (e.g., PostHog, Amplitude)
-  // Example: posthog.capture(event, data);
 }
+
+// ============================================================================
+// HOOK
+// ============================================================================
 
 export interface UseBoardOnboardingOptions {
   projectPath: string | null;
-  isEmpty: boolean; // Whether the board has no features
-  totalFeatureCount: number; // Total number of features in the board
-  /** Whether the spec generation dialog is currently open (prevents wizard from showing) */
-  isSpecDialogOpen?: boolean;
 }
 
 export interface UseBoardOnboardingResult {
-  // Wizard visibility
+  // From shared wizard hook
   isWizardVisible: boolean;
-  shouldShowWizard: boolean;
-
-  // Current step
   currentStep: number;
-  currentStepData: WizardStep | null;
+  currentStepData: OnboardingStep | null;
   totalSteps: number;
-
-  // Navigation
   goToNextStep: () => void;
   goToPreviousStep: () => void;
   goToStep: (step: number) => void;
-
-  // Actions
   startWizard: () => void;
   completeWizard: () => void;
   skipWizard: () => void;
-  dismissWizard: () => void;
+  isCompleted: boolean;
+  isSkipped: boolean;
 
-  // Quick Start / Sample Data
+  // Board-specific
   hasSampleData: boolean;
   setHasSampleData: (has: boolean) => void;
   markQuickStartUsed: () => void;
 
-  // Re-trigger
-  canRetrigger: boolean;
-  retriggerWizard: () => void;
-
-  // State
-  isCompleted: boolean;
-  isSkipped: boolean;
+  // Steps data for component
+  steps: OnboardingStep[];
 }
 
 export function useBoardOnboarding({
   projectPath,
-  isEmpty,
-  totalFeatureCount,
-  isSpecDialogOpen = false,
 }: UseBoardOnboardingOptions): UseBoardOnboardingResult {
-  // Local state
-  const [currentStep, setCurrentStep] = useState(0);
-  const [isWizardActive, setIsWizardActive] = useState(false);
-  const [onboardingState, setOnboardingState] = useState<OnboardingState>(DEFAULT_ONBOARDING_STATE);
+  // Board-specific state for sample data
+  const [boardData, setBoardData] = useState<BoardOnboardingData>(DEFAULT_BOARD_DATA);
 
-  // Load persisted state when project changes
+  // Create storage key from project path
+  const storageKey = projectPath ? `board:${sanitizeProjectPath(projectPath)}` : 'board:default';
+
+  // Use the shared onboarding wizard hook
+  const wizard = useOnboardingWizard({
+    storageKey,
+    steps: BOARD_WIZARD_STEPS,
+  });
+
+  // Load board-specific data when project changes
   useEffect(() => {
     if (!projectPath) {
-      setOnboardingState(DEFAULT_ONBOARDING_STATE);
+      setBoardData(DEFAULT_BOARD_DATA);
       return;
     }
 
-    const state = loadOnboardingState(projectPath);
-    setOnboardingState(state);
+    const data = loadBoardData(projectPath);
+    setBoardData(data);
+  }, [projectPath]);
 
-    // Auto-show wizard for empty boards that haven't seen it
-    // Don't re-trigger if board became empty after having features (edge case)
-    // Don't show if spec dialog is open (for new projects)
-    if (
-      isEmpty &&
-      !state.hasEverSeenWizard &&
-      !state.completed &&
-      !state.skipped &&
-      !isSpecDialogOpen
-    ) {
-      // Small delay to let the board render first
-      const timer = setTimeout(() => {
-        setIsWizardActive(true);
-        trackAnalytics(ONBOARDING_ANALYTICS.STARTED, { projectPath });
-      }, WIZARD_AUTO_SHOW_DELAY_MS);
-      return () => clearTimeout(timer);
-    }
-  }, [projectPath, isEmpty, isSpecDialogOpen]);
-
-  // Update persisted state helper
-  const updateState = useCallback(
-    (updates: Partial<OnboardingState>) => {
+  // Update board data helper
+  const updateBoardData = useCallback(
+    (updates: Partial<BoardOnboardingData>) => {
       if (!projectPath) return;
 
-      setOnboardingState((prev) => {
-        const newState = { ...prev, ...updates };
-        saveOnboardingState(projectPath, newState);
-        return newState;
+      setBoardData((prev) => {
+        const newData = { ...prev, ...updates };
+        saveBoardData(projectPath, newData);
+        return newData;
       });
     },
     [projectPath]
   );
 
-  // Determine if wizard should be visible
-  // Don't show if:
-  // - No project selected
-  // - Already completed or skipped
-  // - Board has features and user has seen wizard before (became empty after deletion)
-  const shouldShowWizard = useMemo(() => {
-    if (!projectPath) return false;
-    if (onboardingState.completed || onboardingState.skipped) return false;
-    if (!isEmpty && onboardingState.hasEverSeenWizard) return false;
-    return isEmpty && !onboardingState.hasEverSeenWizard;
-  }, [projectPath, isEmpty, onboardingState]);
-
-  // Current step data
-  const currentStepData = WIZARD_STEPS[currentStep] || null;
-  const totalSteps = WIZARD_STEPS.length;
-
-  // Navigation handlers
-  const goToNextStep = useCallback(() => {
-    if (currentStep < totalSteps - 1) {
-      const nextStep = currentStep + 1;
-      setCurrentStep(nextStep);
-      trackAnalytics(ONBOARDING_ANALYTICS.STEP_VIEWED, {
-        step: nextStep,
-        stepId: WIZARD_STEPS[nextStep]?.id,
-        projectPath,
-      });
-    }
-  }, [currentStep, totalSteps, projectPath]);
-
-  const goToPreviousStep = useCallback(() => {
-    if (currentStep > 0) {
-      setCurrentStep(currentStep - 1);
-    }
-  }, [currentStep]);
-
-  const goToStep = useCallback(
-    (step: number) => {
-      if (step >= 0 && step < totalSteps) {
-        setCurrentStep(step);
-        trackAnalytics(ONBOARDING_ANALYTICS.STEP_VIEWED, {
-          step,
-          stepId: WIZARD_STEPS[step]?.id,
-          projectPath,
-        });
-      }
-    },
-    [totalSteps, projectPath]
-  );
-
-  // Wizard lifecycle handlers
-  const startWizard = useCallback(() => {
-    setCurrentStep(0);
-    setIsWizardActive(true);
-    updateState({ hasEverSeenWizard: true });
-    trackAnalytics(ONBOARDING_ANALYTICS.STARTED, { projectPath });
-  }, [projectPath, updateState]);
-
-  const completeWizard = useCallback(() => {
-    setIsWizardActive(false);
-    setCurrentStep(0);
-    updateState({
-      completed: true,
-      completedAt: new Date().toISOString(),
-      hasEverSeenWizard: true,
-    });
-    trackAnalytics(ONBOARDING_ANALYTICS.COMPLETED, {
-      projectPath,
-      quickStartUsed: onboardingState.quickStartUsed,
-      totalFeatureCount,
-    });
-  }, [projectPath, updateState, onboardingState.quickStartUsed, totalFeatureCount]);
-
-  const skipWizard = useCallback(() => {
-    setIsWizardActive(false);
-    setCurrentStep(0);
-    updateState({
-      skipped: true,
-      skippedAt: new Date().toISOString(),
-      hasEverSeenWizard: true,
-    });
-    trackAnalytics(ONBOARDING_ANALYTICS.SKIPPED, {
-      projectPath,
-      skippedAtStep: currentStep,
-    });
-  }, [projectPath, currentStep, updateState]);
-
-  const dismissWizard = useCallback(() => {
-    // Same as skip but doesn't mark as "skipped" - just closes the wizard
-    setIsWizardActive(false);
-    updateState({ hasEverSeenWizard: true });
-  }, [updateState]);
-
-  // Quick Start / Sample Data
+  // Sample data handlers
   const setHasSampleData = useCallback(
     (has: boolean) => {
-      updateState({ hasSampleData: has });
+      updateBoardData({ hasSampleData: has });
       if (!has) {
-        trackAnalytics(ONBOARDING_ANALYTICS.SAMPLE_DATA_CLEARED, { projectPath });
+        trackAnalytics(BOARD_ONBOARDING_ANALYTICS.SAMPLE_DATA_CLEARED, { projectPath });
       }
     },
-    [projectPath, updateState]
+    [projectPath, updateBoardData]
   );
 
   const markQuickStartUsed = useCallback(() => {
-    updateState({ quickStartUsed: true, hasSampleData: true });
-    trackAnalytics(ONBOARDING_ANALYTICS.QUICK_START_USED, { projectPath });
-  }, [projectPath, updateState]);
-
-  // Re-trigger wizard - memoized for stable reference
-  const canRetrigger = useMemo(
-    () => onboardingState.completed || onboardingState.skipped,
-    [onboardingState.completed, onboardingState.skipped]
-  );
-
-  const retriggerWizard = useCallback(() => {
-    setCurrentStep(0);
-    setIsWizardActive(true);
-    // Don't reset completion status, just show wizard again
-    trackAnalytics(ONBOARDING_ANALYTICS.RETRIGGERED, { projectPath });
-  }, [projectPath]);
+    updateBoardData({ quickStartUsed: true, hasSampleData: true });
+    trackAnalytics(BOARD_ONBOARDING_ANALYTICS.QUICK_START_USED, { projectPath });
+  }, [projectPath, updateBoardData]);
 
   return {
-    // Visibility
-    isWizardVisible: isWizardActive,
-    shouldShowWizard,
+    // Spread shared wizard state and actions
+    isWizardVisible: wizard.isVisible,
+    currentStep: wizard.currentStep,
+    currentStepData: wizard.currentStepData,
+    totalSteps: wizard.totalSteps,
+    goToNextStep: wizard.goToNextStep,
+    goToPreviousStep: wizard.goToPreviousStep,
+    goToStep: wizard.goToStep,
+    startWizard: wizard.startWizard,
+    completeWizard: wizard.completeWizard,
+    skipWizard: wizard.skipWizard,
+    isCompleted: wizard.isCompleted,
+    isSkipped: wizard.isSkipped,
 
-    // Steps
-    currentStep,
-    currentStepData,
-    totalSteps,
-
-    // Navigation
-    goToNextStep,
-    goToPreviousStep,
-    goToStep,
-
-    // Actions
-    startWizard,
-    completeWizard,
-    skipWizard,
-    dismissWizard,
-
-    // Sample Data
-    hasSampleData: onboardingState.hasSampleData,
+    // Board-specific
+    hasSampleData: boardData.hasSampleData,
     setHasSampleData,
     markQuickStartUsed,
 
-    // Re-trigger
-    canRetrigger,
-    retriggerWizard,
-
-    // State
-    isCompleted: onboardingState.completed,
-    isSkipped: onboardingState.skipped,
+    // Steps data
+    steps: BOARD_WIZARD_STEPS,
   };
 }
